chore: Move the example programs into `examples`
