+++ quizx/examples/compute_tensor.rs
