+++ quizx/examples/gauss.rs
