+++ quizx/examples/hidden_shift_stabrank.rs
