+++ quizx/examples/pauli_gadget_single_q.rs
