+++ quizx/examples/pauli_gadget_stabrank.rs
