+++ quizx/examples/pauli_gadget_stabrank_met.rs
