+++ quizx/examples/pauli_gadget_stabrank_ne.rs
