+++ quizx/examples/pauli_gadget_stabrank_single.rs
