+++ quizx/examples/scratch.rs
