+++ quizx/examples/stabrank.rs
