+++ quizx/examples/test_all.rs
@@ -23,12 +23,12 @@ use std::fs;
 use std::time::Instant;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    for e in fs::read_dir("../circuits/small")? {
+    for e in fs::read_dir("circuits/small")? {
         if let Some(f) = e?.path().to_str() {
             let time = Instant::now();
             println!("{}", f);
-            let c =
-                Circuit::from_file(f).unwrap_or_else(|_| panic!("circuit failed to parse: {}", f));
+            let c = Circuit::from_file(f)
+                .unwrap_or_else(|e| panic!("circuit failed to parse: {}. {}", f, e));
             println!("...done reading in {:.2?}", time.elapsed());
             // if c.num_qubits() > 10 { continue; }
