+++ quizx/examples/test_one.rs
@@ -22,10 +22,11 @@ use std::time::Instant;
 use std::{thread, time};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let f = "../circuits/large/gf2^64_mult.qasm";
+    let f = "circuits/large/gf2^64_mult.qasm";
     let time = Instant::now();
     println!("{}", f);
-    let c = Circuit::from_file(f).unwrap_or_else(|_| panic!("circuit failed to parse: {}", f));
+    let c =
+        Circuit::from_file(f).unwrap_or_else(|e| panic!("circuit failed to parse: {}. {}", f, e));
     println!("...done reading in {:.2?}", time.elapsed());
 
     // println!("Computing tensor");
